Refactor json utility to wxdata, add ReadJsonString function
