+++ scwx-qt/source/scwx/qt/model/layer_model.cpp
@@ -1,7 +1,7 @@
 #include <scwx/qt/model/layer_model.hpp>
 #include <scwx/qt/manager/placefile_manager.hpp>
 #include <scwx/qt/types/qt_types.hpp>
-#include <scwx/qt/util/json.hpp>
+#include <scwx/util/json.hpp>
 #include <scwx/util/logger.hpp>
 
 #include <filesystem>
